# perf: Optimize read MIME operation