**What included in this PR**
- Set FileStream's BufferSize to `0` (Default:4096)
- Set StreamReader's BufferSize to `128` (Default:1024, Min:128)
- Set `FileOptions.SequentialScan` hint for Windows (It seems don't affects actual performance thought)

**Background**
`YamlMime.ReadMime` API read file's first line only.
So it can reduce internal buffer allocations. and reduce Gen0 GC allocations.

**Benchmark results**

|              Method |     Mean |    Error |   StdDev | Ratio |   Gen0 |   Gen1 | Allocated | Alloc Ratio |
|-------------------- |---------:|---------:|---------:|------:|-------:|-------:|----------:|------------:|
|            ReadMime | 19.23 μs | 0.192 μs | 0.180 μs |  1.00 | 1.8616 | 0.0305 |   7.73 KB |        1.00 |
|  ReadMime_Optimized | 18.16 μs | 0.083 μs | 0.065 μs |  0.94 | 0.2441 | 0.0305 |   1.02 KB |        0.13 |